Implement all CRUD operations for Tasks. Closes #3

+++ src/main/java/com/danpopescu/taskmanagement/services/impl/TaskServiceImpl.java
@@ -1,11 +1,11 @@
 package com.danpopescu.taskmanagement.services.impl;
 
 import com.danpopescu.taskmanagement.domain.Task;
+import com.danpopescu.taskmanagement.dto.TaskDTO;
+import com.danpopescu.taskmanagement.exceptions.TaskNotFoundException;
 import com.danpopescu.taskmanagement.repositories.TaskRepository;
 import com.danpopescu.taskmanagement.services.TaskService;
 import org.springframework.stereotype.Service;
-
-import java.util.Optional;
 
 @Service
 public class TaskServiceImpl implements TaskService {
@@ -22,22 +22,44 @@ public class TaskServiceImpl implements TaskService {
     }
 
     @Override
-    public Optional<Task> getById(Long id) {
-        return taskRepository.findById(id);
+    public Task save(TaskDTO taskDTO) {
+        Task task = dtoToTask(taskDTO);
+        return taskRepository.save(task);
+    }
+
+    private Task dtoToTask(TaskDTO taskDTO) {
+        return Task.builder()
+                .id(taskDTO.getId())
+                .title(taskDTO.getTitle())
+                .done(taskDTO.isDone())
+                .build();
     }
 
     @Override
-    public Iterable<Task> getAll() {
+    public Task findById(Long id) throws TaskNotFoundException {
+        return taskRepository.findById(id).orElseThrow(() -> new TaskNotFoundException("Task Not Found"));
+    }
+
+    @Override
+    public Iterable<Task> findAll() {
         return taskRepository.findAll();
     }
 
     @Override
-    public void deleteById(Long id) {
+    public void deleteById(Long id) throws TaskNotFoundException {
+        if (!taskRepository.existsById(id)) {
+            throw new TaskNotFoundException("Task Not Found");
+        }
         taskRepository.deleteById(id);
     }
 
     @Override
     public void deleteAll(Iterable<Task> tasks) {
         taskRepository.deleteAll(tasks);
+    }
+
+    @Override
+    public boolean existsById(Long id) {
+        return taskRepository.existsById(id);
     }
 }
